리뷰 카운트 안 맞는 부분 수정(본인 리뷰 조회, 학원 리뷰 조회), JWT 인증 강화

+++ src/main/java/com/green/acamatch/academy/model/HB/MediaReviewDto.java
@@ -18,7 +18,7 @@ public class MediaReviewDto {
     private int reviewStar;
     private String reviewCreatedAt;
     private String reviewUpdatedAt;
-    private int reviewUserId;
+    private Long reviewUserId;
     private String reviewUserNickName;
     private String reviewClassName;
     private int joinClassId;
